remove code from __init__
