--- src/upload/ut/__init__.py
@@ -1,14 +1,0 @@
-import unittest
-import json
-
-import src.upload.router as router
-
-
-class BasicTest(unittest.TestCase):
-    """Test basic functionality for "upload" route."""
-
-    def test_base(self):
-        """Test basic."""
-        self.assertEqual(router.upload_http("some name")["id"], 1)
-        self.assertEqual(router.upload_http("more name")["id"], 2)
-        self.assertEqual(router.upload_http("new name")["id"], 3)
